feat(techdocs): add extension for setting build strategy
