+++ plugins/techdocs-node/api-report.md
@@ -10,6 +10,7 @@ import { Config } from '@backstage/config';
 import { ContainerRunner } from '@backstage/backend-common';
 import { Entity } from '@backstage/catalog-model';
 import express from 'express';
+import { ExtensionPoint } from '@backstage/backend-plugin-api';
 import { IndexableDocument } from '@backstage/plugin-search-common';
 import { Logger } from 'winston';
 import { PluginEndpointDiscovery } from '@backstage/backend-common';
@@ -22,6 +23,12 @@ export class DirectoryPreparer implements PreparerBase {
   static fromConfig(config: Config, options: PreparerConfig): DirectoryPreparer;
   prepare(entity: Entity, options?: PreparerOptions): Promise<PreparerResponse>;
   shouldCleanPreparedDirectory(): boolean;
+}
+
+// @public
+export interface DocsBuildStrategy {
+  // (undocumented)
+  shouldBuild(params: ShouldBuildParameters): Promise<boolean>;
 }
 
 // @public
@@ -227,7 +234,21 @@ export type ReadinessResponse = {
 export type RemoteProtocol = 'url' | 'dir';
 
 // @public
+export type ShouldBuildParameters = {
+  entity: Entity;
+};
+
+// @public
 export type SupportedGeneratorKey = 'techdocs' | string;
+
+// @public
+export interface TechdocsBuildStrategyExtensionPoint {
+  // (undocumented)
+  setBuildStrategy(buildStrategy: DocsBuildStrategy): void;
+}
+
+// @public
+export const techdocsBuildStrategyExtensionPoint: ExtensionPoint<TechdocsBuildStrategyExtensionPoint>;
 
 // @public
 export interface TechDocsDocument extends IndexableDocument {
