+++ plugins/techdocs-node/src/extensions.ts
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2023 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { createExtensionPoint } from '@backstage/backend-plugin-api';
+import { DocsBuildStrategy } from './techdocsTypes';
+
+/**
+ * Extension point type for setting a custom build strategy.
+ *
+ * @public
+ */
+export interface TechdocsBuildStrategyExtensionPoint {
+  setBuildStrategy(buildStrategy: DocsBuildStrategy): void;
+}
+
+/**
+ * Extension point for setting a custom build strategy.
+ *
+ * @public
+ */
+export const techdocsBuildStrategyExtensionPoint =
+  createExtensionPoint<TechdocsBuildStrategyExtensionPoint>({
+    id: 'techdocs.buildStrategy',
+  });
