+++ plugins/techdocs-node/src/techdocsTypes.ts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+import { Entity } from '@backstage/catalog-model';
 import { IndexableDocument } from '@backstage/plugin-search-common';
 
 /**
@@ -45,4 +46,22 @@ export interface TechDocsDocument extends IndexableDocument {
    * Entity path
    */
   path: string;
+}
+
+/**
+ * Parameters passed to the shouldBuild method on the DocsBuildStrategy interface
+ *
+ * @public
+ */
+export type ShouldBuildParameters = {
+  entity: Entity;
+};
+
+/**
+ * A strategy for when to build TechDocs locally, and when to skip building TechDocs (allowing for an external build)
+ *
+ * @public
+ */
+export interface DocsBuildStrategy {
+  shouldBuild(params: ShouldBuildParameters): Promise<boolean>;
 }
